fix(deps): update rust crates libsqlite3-sys to 0.36.0 and rusqlite to 0.38.0 (#91)

+++ src/command/snapshot/add_server.rs
@@ -63,7 +63,7 @@ impl AddServerCommand {
                 VALUES (:id, :address, :role);
             ",
             named_params! {
-                ":id": id,
+                ":id": id as i64,
                 ":address": address,
                 ":role": role as u8,
             },
